fixes and improvements
